requirements.txt
requests>=2.32.3
beautifulsoup4>=4.12.3
pandas>=2.2.2
openpyxl>=3.1.5
xlsxwriter>=3.2.0
matplotlib>=3.9.0
seaborn>=0.13.2
lxml>=5.2.2
python-dateutil>=2.9.0.post0
retrying>=1.3.4